+++ literary_system/learning/__init__.py
@@ -17,3 +17,5 @@ from literary_system.learning.rlaif_orchestrator import (
 from literary_system.learning.rlaif_trigger import TriggerResult, RLAIFTrigger
 
 from .phase_e_exit import run_phase_e_exit  # V766
+
+from .provider_router import ProviderRouter, RoutingSignals, RoutingDecision, validate_routing  # V768